Allow to filter by ref attribute values
This needs special treatment since we'll want to query for specific
references via e.g.

   (... {:user/friends <id>})

where :user/friends actual stores a collection of references. In this
case matching against the ID is a member check, not an equality check.
Furthermore, the references may be stored in maps with :db/id
attributes:

   {... :user/friends [{:db/id <id1>} {:db/id <id2>}]}

In this case we'll have to extract the IDs first and then compare
the value against them.