Only run a single pull-based query when fetching all entities
This is instead of first running one query to obtain only the IDs
of the entities and then running the regular fetch-entities function
to execute pull queries for all these IDs. With DataScript, it's faster
if we do both in one go.

Also move the has-entity-attrs? rule generating code into util to share
it between the regular DataScript data layer and the data layer without
authorization.