Add :skip-authorization? env option to Datomic/DataScript data layers
This allows to bypass authorization entirely when accessing data
through the Datomic and DataScript data layers. It's a prerequisite
for running entity auth queries, which have to be processed without
authorization in order to have unrestricted access to the database.

When running auth queries with the query engine, we have to set this
flag in order to prevent infinite recursion (caused by performing
authorization on auth queries, which would then trigger another
auth query etc. etc.).