Run auth queries in query tests without a cache
This is currently necessary, as multiple sub-queries of an auth
query might query for the same data but with different parameters.

If we use a cache, then the first of such sub-queries will put
the data into the cache if it matches the params. Later sub-queries
for the same data would not re-match the data against the params
and may thus - incorrectly - return the data.

In fact, this is a problem for all queries, not just auth queries, so
we will have to rethink our approach to caching (e.g. by re-authorizing
and re-matching items that are already cached, or by dropping the cache
concept entirely).