Add a :ref-id-attr option to the query engine environment
This can be used to express which attribute is used as the default
ID attribute in references by the data layer. E.g. Datomic and
DataScript both store references as {:db/id <id>} maps. Here, the
:ref-id-attr would be :db/id.

The map-based data layer stores references as {<id-attr> <id>}, where
<id-attr> is derived from the :id-attr environment option (in our
case that is usually set to :workflo/id).