Fix fetching all entities in the Datomic data layer
When fetching all instances of an entity, we first query all IDs and
then carry on as if we were fetching many instances with specific IDs.
When we do that, we need to ensure that the ref-id-attr matches the
id-attr, so that the IDs returned from the first part of the logic
match the ref-id-attr that is used in the second part of the logic.

This could all be reduced to a single code path if the Datomic data
layer were changed to expand attributes like :some/reference to
{:some-reference [id-attr]} so that we can always operate against
id-attr everywhere instead of having to deal with a mix of :db/id
and :workflo/id.